Correct misplaced doc strings and deftest duplicate name
The change from ^Map to ^java.util.Map type tag is not needed for
Clojure, nor is it harmful for Clojure, but it does help the
pre-release version of the Eastwood Clojure lint tool to avoid
throwing an exception (the cause of which is being investigated at the
Clojure 'spec' level with Clojure developers).

The deftest that had a duplicate name was causing some unit tests
never to run.  All tests still pass after the rename to make them
unique, at least on my dev environment.