Fix #24, support vector as property value
As described in #24, we were unable to store vectors as values in TinkerPop because of flatten.

This replaces the IMO unnecessary `flatten` with the much more, single-level `apply concat`.

I decided to keep the old code for non-map inputs (as those some tests provide) to keep backwards compatibility with any clients using those.